k22f: Remove unused MK22F10.h header
According to NXP material, there are no 100 MHz K22F parts in 144 pin
packages, which is the reference manual that this header is supposed to
correspond to.
The header was originally included from Keil uVision Kinetis support
packages. It is possible that this header was only used with
engineering samples of the K22F.