tests: deactivate DEVELHELP for unittests
In my opinion it doesn't make sense to activate DEVELHELP for unittests
because it also causes the evaluation of `assert()`. This causes the
tests to crash for values the tests test intentionally (because it is
the point of the unittests to test also some unexpected values and how
this influences the behavior).