dhcpv6: provide initial client implementation
This implements a client for DHCPv6 IA_PD (Identity Association for
Prefix Delegation). Goal was to have a IETF-compliant alternative to
UHCP. The implementation was based on draft-ietf-dhc-rfc3315bis-12.

We currently do not have a stack-independent API to access network
interfaces, so every configuration of those is still GNRC-dependent,
though the implementation is to be targeted to be stack independent one
day.