boards/iotlab-m3|a8: fix openocd configuration
iotlab-m3 boards always ended up not being able to flash after time.
This changes managed to fix and flash boards that where able to be flashed with
the old `ftdi2232` driver and not `RIOT`

It combines configuration from openocd, iot-lab and RIOT config

 * http://repo.or.cz/openocd.git/blob/HEAD:/tcl/interface/ftdi/iotlab-usb.cfg
   * ftdi configuration
 * https://github.com/iot-lab/iot-lab-gateway/blob/2.4.1/gateway_code/static/iot-lab-m3.cfg
   * `trst_and_srst` config
 * RIOT
   * Kee the `configure -rtos` auto

I do not understand the configuration, just adapted and tested.