makefiles/docker.inc.mk: allow using sudo for docker
Add `DOCKER_WITH_SUDO` variable that runs docker using `sudo` when set to 1.

Not all users register their user in the docker group and should not run the
whole make process as root to use docker.
Creating files as a user is correctly handled as `id -u` is still run as the
original user.