pkg/libfixmath: Replace __FILE__ by RIOT_FILE_NOPATH
The binary file size changes when the RIOT directory is moved.
This caused the `libfixmath_unittests` to fail on my computer.

I used RIOT_FILE_NOPATH instead of RIOT_FILE_RELATIVE as 'TEST' is used a lot
and it would allow more boards to be tested, full path is not that important.