gnrc_ipv6: fix regression from #8215
While `tmp` in the loop for write-protection for the check-sum
calculation is used to check the return value of
`gnrc_pktbuf_start_write()`, it never overwrites `payload` causing the
original snip to be used in the following iteration `prev` when
duplicated, and destroying the sanity of `ipv6`. However, since no
release in this function is required anyways, we can just use `payload`
to check the return value directly.