tests/libc_newlib: fix pointers comparison for llvm
With llvm and samr21-xpro, I could not directly do 'printf == iprintf'.
And I could not cast them to an integer value in a portable way.
So I use real function pointers for this.
However comparing `printf_addr` with `iprintf_addr` does not work if
there is not the memcmp to somehow 'force them' to be pointers...
So I used the 'memcmp' for the comparison.