tests/tweetnacl: Move from unittests to regular test
Stack size is changed from 4 times the default + printf to 5 times the
default stack size. Only on the lpc2387 this reduces the resulting stack
space. the test is not rerun for the lpc2387 and is untested.

Tests are disable for the Arduino. While they "might" work, it takes
around 8 minutes to complete the tests.