sys/stdio_uart: do not cast function pointer.
There was a function pointer cast in stdio_uart. The warning was
being explicitly supressed in the module's makefile.
This commit removes the suppression and fixes the warning by providing
a wrapper function.
The wrapper was already defined in the AT driver, as part of a fix for
a similar bug there, so the definition was moved to the isrpipe module,
the logic being that using isrpipe in conjunction with a periph uart
is a common enough use-case.