boards/msba2: Do not set PORT.
Boards should not set PORT and should not have code conditional on
PORT as that causes PORT to be evaluated and the build to fail even
if this varible is not needed.

Exporting has the same effect.

This fixes the MSBA2 board by declaring PORT_LINUX and PORT_DARWIN
instead. PORT_DARWIN is obviously wrong, but it is broken with or
without this PR.