Makefile.include: require make version 4.
Old versions of GNU Make (especially on OSX) are a pain to deal
with. This commit introduces a warning when such an old version
is found. The goal is the future the warning will be turned into
an error.

Complying with version 4.x is easy on OSX by using homebrew, and
on Linux, even Debian stable has the required version.