drivers/ds18: refactor functions with const dev pointer
All DS18 functions have a dev argument. All except the init function use
it as an IN parameter, so we can prototype it as const ds18_t*.

As a consequence we can drop the cast in read_temperature() in ds18_saul.c
which was the primary trigger for the changes.

The commit also follows the preferred convention that "params" is a field
in the device struct. Only the init function needs to write it.