makefiles/toolchain: fallback to 'objdump'
When '$(PREFIX)objdump' is not present fallback to 'objdump'.
'objdump' is used when flashing for some boards but the toolchain may
not be installed when building in docker.

If none is found, it defaults to saying 'false' but not sure if it is a
decent solution or not. Maybe just always default to 'objdump'?

This will allow using 'objdump' in 'cpu/kinetis/dist/check-fcfield-elf.sh'.