tests/test_tools: add a test for the testing tools environment
It checks the interraction with a board through make term.
It is using a shell without echo or prompt.

It currently checks that it is not adding unwanted output:

* local echo
* getting some test output without other messages
    * This one currently does not work

HACK: pyterm is currently adding timestamp and garbage output while
testing so it needs special handling to try matching only an integer.