makefiles: introduce rawterm recipe and variables.
`make rawterm` is the new interface for automated access to the serial
port. It will be used in testing. In contrast with `make term`, which is
intended for human-machine-interaction, rawterm is meant to comminicate
the node with automated scripts running on the host machine.

rawterm must:

- Not have any input or output buffering.
- Not modify input or output in any way (no extra prompts, logging artifacts,
  translations etc.)
- Have local echo disabled.
- Not print anything to stdout that was not generated by the node (i.e. no
  startup banners.)

With this definition it is very easy to implement a generic "nice" terminal
(that is `make term`) as a wrapper around rawterm.