tests/mcuboot: handle building 'mcuboot' in docker
When using BUILD_IN_DOCKER only make 'all' execute 'mcuboot' target when
inside the container. Otherwise it will execute it using your host
toolchain.

This is a hack and could be prevented by having a file target put in
FLASHFILE.

This does currently not allow flashing after building in docker as
flashing should be done with `mcuboot-flash` which will not be done in
docker. It would require the `FLASHFILE` fix to be done properly.