testrunner: Use `make rawterm`
Use the raw terminal to run the tests. This is supposed to simplify the
test scripts as it it no longer possible for the output/decorations of
the terminal program to be confused with the output of the node. In
addition, the raw terminal is more likely to be more lightweight.

The lack of buffering in the raw terminal makes it possible to test low
level uart functions.

Decoupling the UI terminal from the testing terminal also gives us more
freedom to improve or change the UI without worrying about breaking the
tests.