tests/gnrc_udp: Replace atoi() by strtol().
atoi() cannot detect errors. Many implementation return zero on error
and that is what was being checked here, making the "udp send" command
unable to parse integer values of zero. On top of this, the behavior on
errors does not seem to be specified in the standard (so it is not even
correct to check for zero even when zero is not an accepted value, like
for a port number).

The result of all this is that sending UDP packets of zero length (as
required by the Release Specs) was not possible.

This patch replaces atoi by strlen, which allows for robust error detection.
Sending zero length packets is possible.