cortexm_common/cpu_check_address: fixup for asm statement.
To prevent GCC from optimizing away what it should not, the best way is
to use Inputs, Outputs and Clobbers to tell it exacly what we are doing.

With this change the code works again in cortex-m0.

For more information see:

https://gcc.gnu.org/onlinedocs/gcc/Extended-Asm.html#Extended-Asm

Objdump show the following asm being generated (gcc):

-------------------------------
00000000 <cpu_check_address>:

bool cpu_check_address(volatile const char *address)
{
   0:	b520      	push	{r5, lr}
    /* Cortex-M0 doesn't have BusFault so we need to catch HardFault */
    (void)address;

    bool result;

    __asm__ volatile (
   2:	2501      	movs	r5, #1
   4:	4902      	ldr	r1, [pc, #8]	; (10 <cpu_check_address+0x10>)
   6:	4a03      	ldr	r2, [pc, #12]	; (14 <cpu_check_address+0x14>)
   8:	7803      	ldrb	r3, [r0, #0]
   a:	1c28      	adds	r0, r5, #0
   c:	b2c0      	uxtb	r0, r0
        : "r1", "r2", "r3", "r5"
     );

    return result;
}
   e:	bd20      	pop	{r5, pc}
  10:	deadf00d 	.word	0xdeadf00d
  14:	cafebabe 	.word	0xcafebabe
-----------------------------

In this particular case clang seems to be a bit less smart and also uses
r4 and r7.