stm32f103c8: handle default STM32F103C8_FLASH_HACK without variable
Handle the `STM32F103C8_FLASH_HACK` variable handling without having to
set it to 0 by default.

This is part of cleanup prior to moving the CPU/CPU_MODEL to Makefile.features.
It is better if BSP specific variables can be avoided or would need
otherwise special handling when calculating 'info-boards-supported'.