riotboot: set FLASHFILE to RIOTBOOT_EXTENDED_BIN
... if the riotboot feature is used.

Previously, even an application that had "FEATURES_REQUIRED += riotboot"
set would still flash the non-riotboot binary on "make flash".
This is usualy not what the user wants.

This commit set's the FLASHFILE variable to the combined "riotboot
bootloader + slot0 + empty slot1" binary. This has the effect that make
fall, flash and flash-only will compile and/or flash a working riotboot
setup.